# PERF: parse and timedelta ops improvements, #6755