@jreback 2015-06-20
This is a cython conversion of the timedelta parsing routines. Was much more involved that it looked, as the original was using regexes, this is a string parsing state machine in cython.

However, I think this could actually be done in c (or maybe using ``char *`` pointers in cython), rather than the list build up. This could be another 10x perf improvement on this (my benchmark is the c parser for ISO datetimes).

if any of the c gurus are interested, pls comment!